requirements.txt
opencv-python>=4.8.0
pyzbar>=0.1.9
numpy>=1.24.0

# NFC支持 (可选)
nfcpy>=1.0.4

# 键盘模拟 (可选，用于实际翻页操作)
pynput>=1.7.6

# 串口通信 (可选，用于某些NFC读卡器)
pyserial>=3.5 